+++ src/main/java/tornadofx/CSS.kt
@@ -1,5 +1,9 @@
 package tornadofx
 
+import javafx.beans.value.ChangeListener
+import javafx.beans.value.ObservableValue
+import javafx.scene.Node
+
 import javafx.scene.paint.Color
 import javafx.scene.paint.Paint
 import java.text.DecimalFormat
@@ -507,3 +511,25 @@ enum class FXBarPolicy(val value: String) {
 enum class FXHAlignment { left, center, right; }
 enum class FXVAlignment { top, center, baseline, bottom; }
 enum class FXTabAnimation { grow, none; }
+
+class ObservableStyleClass(node: Node, val value: ObservableValue<String>) {
+    val listener: ChangeListener<String>
+
+    init {
+        fun checkAdd(newValue: String?) {
+            if (newValue != null && !node.hasClass(newValue)) node.addClass(newValue)
+        }
+
+        listener = ChangeListener { observableValue, oldValue, newValue ->
+            if (oldValue != null) node.removeClass(oldValue)
+            checkAdd(newValue)
+        }
+
+        checkAdd(value.value)
+        value.addListener(listener)
+    }
+
+    fun dispose() = value.removeListener(listener)
+}
+
+fun Node.bindClass(value: ObservableValue<String>): ObservableStyleClass = ObservableStyleClass(this, value)